feat(cluster): add ScyllaCluster and ClusterConfig structs feat(cluster): add constructor to ScyllaCluster feat(cluster): add connect method to ScyllaCluster feat(session): add ScyllaSession struct feat(session): add execute method to ScyllaSession
This commit adds support for ScyllaCluster and ScyllaSession structs. ScyllaCluster has a constructor that takes a ClusterConfig object and a connect method that returns a ScyllaSession object. ScyllaSession has an execute method that takes a query string and returns a serde_json::Value object.
