+++ src/cluster/cluster_config.rs
@@ -1,0 +1,4 @@
+#[napi(object)]
+pub struct ClusterConfig {
+  pub nodes: Vec<String>,
+}
